move supported format definitions to separate crate
