+++ Cargo.toml
@@ -36,6 +36,7 @@ plotinator-plot-util.path = "crates/plotinator-plot-util"
 plotinator-hdf5.path = "crates/plotinator-hdf5"
 plotinator-mqtt.path = "crates/plotinator-mqtt"
 plotinator-macros.path = "crates/plotinator-macros"
+plotinator-supported-formats.path = "crates/plotinator-supported-formats"
 egui = { version = "0.31", features = ["rayon", "callstack"] }
 egui_plot = { version = "0.31", features = ["serde"] }
 log = "0.4"
@@ -75,6 +76,7 @@ plotinator-plot-util.workspace = true
 plotinator-hdf5.workspace = true
 plotinator-test-util.workspace = true
 plotinator-macros.workspace = true
+plotinator-supported-formats.workspace = true
 egui_plot.workspace = true
 log.workspace = true
 serde.workspace = true
@@ -110,7 +112,7 @@ egui_kittest = { version = "0.31.1", features = ["eframe", "snapshot", "wgpu", "
 env_logger = "0.11"
 zip = "4.0.0"
 axoupdater = { version = "0.9.0", features = ["blocking"], optional = true}
-plotinator-mqtt = { workspace = true, optional = true}
+plotinator-mqtt = { workspace = true, optional = true }
 profiling = { workspace = true, optional = true}
 puffin = { workspace = true, optional = true}
 puffin_http = { workspace = true, optional = true}
@@ -130,7 +132,7 @@ elevated-command.workspace = true
 
 [features]
 default = ["selfupdater", "mqtt", "hdf5"]
-hdf5 = []
+hdf5 = ["plotinator-supported-formats/hdf5"]
 selfupdater = ["dep:axoupdater"]
 mqtt = ["dep:plotinator-mqtt"]
 profiling = ["dep:profiling", "dep:puffin", "dep:puffin_http"]
