+++ crates/plotinator-supported-formats/src/hdf5.rs
