+++ crates/plotinator-supported-formats/src/logs/parse_info.rs
