+++ src/app.rs
@@ -9,13 +9,12 @@ use plotinator_log_if::prelude::Plotable as _;
 
 use file_dialog as fd;
 use loaded_files::LoadedFiles;
-use supported_format::SupportedFormat;
+use plotinator_supported_formats::SupportedFormat;
 
 mod dropped_files;
 mod file_dialog;
 
 pub mod loaded_files;
-pub mod supported_format;
 mod util;
 
 /// if a log is loaded from content that exceeds this many unparsed bytes:
