+++ src/plot/util.rs
@@ -1,8 +1,7 @@
 use egui::{Modifiers, Vec2};
 use plotinator_log_if::prelude::*;
 use plotinator_plot_util::{Plots, StoredPlotLabels};
-
-use crate::app::supported_format::SupportedFormat;
+use plotinator_supported_formats::SupportedFormat;
 
 use super::plot_settings::{PlotSettings, date_settings::LoadedLogSettings};
